The Sampler can now return a random action, generated with the help of the ObjectGenes (rather than just randomizing the other types of genes).
Gene selection has also been updated.

Includes minor refactoring and other corrections (including minor bugfixing for lcs calculation).
