+++ experiments/src/main/kotlin/org/evomaster/experiments/objects/service/ObjRestSampler.kt
@@ -18,7 +18,8 @@ import org.evomaster.core.remote.service.RemoteController
 import org.evomaster.core.search.Action
 import org.evomaster.core.search.gene.*
 import org.evomaster.core.search.service.Sampler
-import org.evomaster.experiments.objects.*
+import org.evomaster.experiments.objects.ObjRestCallAction
+import org.evomaster.experiments.objects.ObjRestActionBuilder
 import org.evomaster.experiments.objects.RestPath
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -28,7 +29,6 @@ import javax.ws.rs.client.ClientBuilder
 import javax.ws.rs.core.MediaType
 import javax.ws.rs.core.Response
 import org.evomaster.experiments.objects.param.PathParam
-
 
 class ObjRestSampler : Sampler<ObjIndividual>() {
 
@@ -50,12 +50,6 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
     private var sqlInsertBuilder: SqlInsertBuilder? = null
 
     private val modelCluster: MutableMap<String, ObjectGene> = mutableMapOf()
-
-    private val uo: UsedObj = UsedObj()
-    //private val usedObj: MutableMap<Pair<String, String> , ObjectGene> = mutableMapOf()
-    //private val usedObj: MutableMap<Pair<ObjRestCallAction, String> , ObjectGene> = mutableMapOf()
-
-    //private val usedObj: MutableList<ObjectGene> = mutableListOf()
 
     @PostConstruct
     private fun initialize() {
@@ -196,95 +190,16 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
         val n = randomness.nextInt(1, config.maxTestSize)
 
         (0 until n).forEach {
-            //actions.add(sampleRandomAction(0.05))
-            actions.add(sampleRandomObjCallAction(0.05))
+            actions.add(sampleRandomAction(0.05))
         }
 
-        val objIndivid = ObjIndividual(actions, SampleType.RANDOM, uo)
-        //objIndivid.uo = uo
-        return objIndivid
+        return ObjIndividual(actions, SampleType.RANDOM)
     }
 
 
     private fun randomizeActionGenes(action: Action) {
-        action.seeGenes().forEach {g ->
-
-            val restrictedModels = mutableMapOf<String, ObjectGene>()
-            modelCluster.forEach{ k, model ->
-                val fields = model.fields.filter { field ->
-                    when (g::class) {
-                        DisruptiveGene::class -> (field as OptionalGene).gene::class === (g as DisruptiveGene<*>).gene::class
-                        OptionalGene::class -> (field as OptionalGene).gene::class === (g as OptionalGene).gene::class
-                        else -> {
-                            false
-                        }
-                    }
-                }
-                //(fields as ArrayList).add(OptionalGene("Deleted", BooleanGene("Deleted", value = false)))
-                restrictedModels[k] = ObjectGene(model.name, fields)
-
-            }
-
-
-            val likely = likelyhoodsExtended(g.getVariableName(), restrictedModels).toList().sortedBy { (_, value) -> -value}.toMap()
-            val sel = pickWithProbability(likely as MutableMap<Pair<String, String>, Float>)
-            val temp = modelCluster.get(sel.first) as ObjectGene
-            temp.randomize(randomness, forceNewValue = true)
-
-            //BMR-update: reversed the order-objects get mutated and I must find a way to propagate the changes back.
-
-            val selectedGene = temp.fields.filter { g -> (g as OptionalGene).name === sel.second }?.single() as OptionalGene
-
-            when (g::class) {
-                DisruptiveGene::class -> (g as DisruptiveGene<*>).gene.copyValueFrom(selectedGene.gene)
-                OptionalGene::class ->  (g as OptionalGene).gene.copyValueFrom(selectedGene.gene)
-            }
-            uo.assign(Pair((action as ObjRestCallAction), g), temp, sel)
-            //TODO: this is where the usedObj is set. Just FYI until the usedObj usage is sorted
-        }
-        uo.coherenceCheck()
+        action.seeGenes().forEach { it.randomize(randomness, false) }
     }
-
-    fun collectObjects(action: Action){
-        /*
-            BMR: ad hoc individuals do not have the same generation path as others
-            as a result, they do not have associated used objects lists
-            this is meant to address that particular problem
-        */
-        action.seeGenes().forEach {g ->
-
-            val restrictedModels = mutableMapOf<String, ObjectGene>()
-            modelCluster.forEach{ k, model ->
-                val fields = model.fields.filter { field ->
-                    when (g::class) {
-                        DisruptiveGene::class -> (field as OptionalGene).gene::class === (g as DisruptiveGene<*>).gene::class
-                        OptionalGene::class -> (field as OptionalGene).gene::class === (g as OptionalGene).gene::class
-                        else -> {
-                            false
-                        }
-                    }
-                }
-                //(fields as ArrayList).add(OptionalGene("Deleted", BooleanGene("Deleted", value = false)))
-                restrictedModels[k] = ObjectGene(model.name, fields)
-
-            }
-
-
-            val likely = likelyhoodsExtended(g.getVariableName(), restrictedModels).toList().sortedBy { (_, value) -> -value}.toMap()
-            val sel = pickWithProbability(likely as MutableMap<Pair<String, String>, Float>)
-            val temp = modelCluster.get(sel.first) as ObjectGene
-            val selectedGene = temp.fields.filter { g -> (g as OptionalGene).name === sel.second }?.single() as OptionalGene
-
-            when (g::class) {
-                DisruptiveGene::class -> (g as DisruptiveGene<*>).gene.copyValueFrom(selectedGene.gene)
-                OptionalGene::class ->  (g as OptionalGene).gene.copyValueFrom(selectedGene.gene)
-            }
-            uo.assign(Pair((action as ObjRestCallAction), g), temp, sel)
-            //TODO: this is where the usedObj is set. Just FYI until the usedObj usage is sorted
-        }
-        uo.coherenceCheck()
-    }
-
 
 
     fun sampleRandomAction(noAuthP: Double): RestAction {
@@ -322,10 +237,9 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
         /*
             At the beginning, sample from this set, until it is empty
          */
-
         if (!adHocInitialIndividuals.isEmpty()) {
             val action = adHocInitialIndividuals.removeAt(adHocInitialIndividuals.size - 1)
-            return ObjIndividual(mutableListOf(action), SampleType.SMART, uo)
+            return ObjIndividual(mutableListOf(action), SampleType.SMART)
         }
 
         if (config.maxTestSize <= 1) {
@@ -340,9 +254,8 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
 
 
         val test = mutableListOf<RestAction>()
-        uo.clearLists()
 
-        val action = sampleRandomObjCallAction(0.0)
+        val action = sampleRandomCallAction(0.0)
 
         /*
             TODO: each of these "smart" tests could end with a GET, to make
@@ -364,7 +277,7 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
         }
 
         if (!test.isEmpty()) {
-            return ObjIndividual(test, sampleType, uo)
+            return ObjIndividual(test, sampleType)
         }
 
         return sampleAtRandom()
@@ -604,7 +517,6 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
         randomizeActionGenes(res)
         res.auth = target.auth
         res.bindToSamePathResolution(target)
-        //TODO this is a candidate for a place to synch objects with paths between related actions?
 
         return res
     }
@@ -681,17 +593,72 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
                 }
     }
 
+    fun getRandomObject(): ObjectGene{
+        val someName = StringGene(
+                "name"
+        )
+        val someId = LongGene(
+                "someId"
+        )
+        val col = mutableListOf(StringGene("a1"), StringGene("a2"), StringGene("a3"))
+        val someCol = MapGene(
+                "map",
+                StringGene("bar"),
+                5,
+                col
+        )
+        someName.randomize(randomness, true)
+        someId.randomize(randomness, true)
+        someCol.randomize(randomness, true)
+        val someFields = listOf(someId, someName, someCol)
+        val generatedObject = ObjectGene(
+                "randomObject",
+                someFields
+        )
+        return generatedObject
+    }
+
+    fun getRandomSwaggerObject(): ObjectGene {
+        var genObj = randomness.choose(modelCluster).copy() as ObjectGene
+        genObj.randomize(randomness, true)
+        return genObj
+    }
+
+    fun getObjectFromTemplate(objTemplate: String): ObjectGene {
+        val result = modelCluster[objTemplate] as ObjectGene
+        result.randomize(randomness, true)
+        return result
+    }
+
+    fun getCompleteObjectFromTemplate(objTemplate: String): ObjectGene {
+        val result = modelCluster[objTemplate] as ObjectGene
+
+        result.randomize(randomness, true)
+        result.fields.forEach{
+            (it as OptionalGene).isActive = true
+        }
+        return result
+    }
+
     fun getModelCluster() :MutableMap<String, ObjectGene>{
         //TODO remove this horrible hack
         return modelCluster
     }
 
-    fun sampleRandomObjCallAction(noAuthP: Double): ObjRestCallAction {
-        val action = randomness.choose(actionCluster.filter { a -> a.value is ObjRestCallAction }).copy() as ObjRestCallAction
-        randomizeActionGenes(action)
-        action.auth = getRandomAuth(noAuthP)
+    fun likelyhoods(parameter: String, candidates: MutableMap<String, ObjectGene>): MutableMap<String, Float>{
 
-        return action
+        val result = mutableMapOf<String, Float>()
+        var sum : Float = 0.toFloat()
+        candidates.forEach { k, v ->
+            val temp = lcs(parameter, k).length.toFloat()/ Integer.max(parameter.length, k.length).toFloat()
+            result[k] = temp
+            sum += temp
+        }
+        result.forEach { k, u ->
+            result[k] = u/sum
+        }
+
+        return result
     }
 
     fun lcs(a: String, b: String): String {
@@ -724,6 +691,57 @@ class ObjRestSampler : Sampler<ObjIndividual>() {
         return found
     }
 
+    fun sampleRandomObjCallAction(noAuthP: Double): ObjRestCallAction {
+        val action = randomness.choose(actionCluster.filter { a -> a.value is ObjRestCallAction }).copy() as ObjRestCallAction
+        action.seeGenes().forEach {g ->
+
+            println(" --- ")
+
+            val restrictedModels = mutableMapOf<String, ObjectGene>()
+            modelCluster.forEach{ k, model ->
+                val fields = model.fields.filter { field ->
+                    when (g::class) {
+                        DisruptiveGene::class -> (field as OptionalGene).gene::class === (g as DisruptiveGene<*>).gene::class
+                        OptionalGene::class -> (field as OptionalGene).gene::class === (g as OptionalGene).gene::class
+                        else -> {
+                            false
+                        }
+                    }
+                }
+
+                restrictedModels[k] = ObjectGene(model.name, fields)
+
+            }
+
+
+            val likely = likelyhoodsExtended(g.getVariableName(), restrictedModels).toList().sortedBy { (_, value) -> -value}.toMap()
+
+            println("Likelyhoods: ${likely}")
+
+
+            println(" -> ${g.getVariableName()}")
+            val sel = pickWithProbability(likely as MutableMap<Pair<String, String>, Float>)
+
+            println("Selected: ${sel.first} -> ${sel.second}")
+
+
+            val temp = modelCluster.get(sel.first) as ObjectGene
+            temp.randomize(randomness, forceNewValue = true)
+
+            val selectedGene = temp.fields.filter { g -> (g as OptionalGene).name === sel.second }?.single() as OptionalGene
+
+
+            when (g::class) {
+                DisruptiveGene::class -> (g as DisruptiveGene<*>).gene.copyValueFrom(selectedGene.gene)
+                OptionalGene::class ->  (g as OptionalGene).gene.copyValueFrom(selectedGene.gene)
+            }
+        }
+
+
+        action.auth = getRandomAuth(noAuthP)
+
+        return action
+    }
 
     fun <T> likelyhoodsExtended(parameter: String, candidates: MutableMap<String, T>): MutableMap<Pair<String, String>, Float>{
